+++ lib/model/db/db_sqflite.dart
@@ -212,7 +212,7 @@ class SqfliteLocalMediaDb implements LocalMediaDb {
     );
     final duplicates = rows.map(AvesEntry.fromMap).toSet();
     if (duplicates.isNotEmpty) {
-      debugPrint('Found duplicates=$duplicates');
+      debugPrint('$runtimeType found duplicates=$duplicates');
     }
     // returns most recent duplicate for each duplicated content ID
     return duplicates;
@@ -507,7 +507,7 @@ class SqfliteLocalMediaDb implements LocalMediaDb {
       if (row != null) {
         result.add(row);
       } else {
-        debugPrint('failed to deserialize cover from row=$rowMap');
+        debugPrint('$runtimeType failed to deserialize cover from row=$rowMap');
       }
     }
     return result;
@@ -565,22 +565,39 @@ class SqfliteLocalMediaDb implements LocalMediaDb {
   // dynamic albums
 
   @override
-  Future<void> clearDynamicAlbums() async {
+  Future<int> clearDynamicAlbums() async {
     final count = await _db.delete(dynamicAlbumTable, where: '1');
     debugPrint('$runtimeType clearDynamicAlbums deleted $count rows');
+    return count;
   }
 
   @override
-  Future<Set<DynamicAlbumRow>> loadAllDynamicAlbums() async {
+  Future<Set<DynamicAlbumRow>> loadAllDynamicAlbums({int bufferSize = _queryCursorBufferSize}) async {
     final result = <DynamicAlbumRow>{};
-    final cursor = await _db.queryCursor(dynamicAlbumTable, bufferSize: _queryCursorBufferSize);
-    while (await cursor.moveNext()) {
-      final rowMap = cursor.current;
-      final row = DynamicAlbumRow.fromMap(rowMap);
-      if (row != null) {
-        result.add(row);
-      } else {
-        debugPrint('failed to deserialize dynamic album from row=$rowMap');
+    try {
+      final cursor = await _db.queryCursor(dynamicAlbumTable, bufferSize: bufferSize);
+      while (await cursor.moveNext()) {
+        final rowMap = cursor.current;
+        final row = DynamicAlbumRow.fromMap(rowMap);
+        if (row != null) {
+          result.add(row);
+        } else {
+          debugPrint('$runtimeType failed to deserialize dynamic album from row=$rowMap');
+        }
+      }
+    } catch (error, stack) {
+      debugPrint('$runtimeType failed to query table=$dynamicAlbumTable error=$error\n$stack');
+      if (bufferSize > 1) {
+        // a large row may prevent reading from the table because of cursor window size limit,
+        // so we retry without buffer to read as many rows as we can, and removing the others
+        debugPrint('$runtimeType retry to query table=$dynamicAlbumTable with no cursor buffer');
+        final safeRows = await loadAllDynamicAlbums(bufferSize: 1);
+        final clearedCount = await clearDynamicAlbums();
+        await addDynamicAlbums(safeRows);
+        final addedCount = safeRows.length;
+        final lostCount = clearedCount - addedCount;
+        debugPrint('$runtimeType kept $addedCount rows, lost $lostCount rows from table=$dynamicAlbumTable');
+        return safeRows;
       }
     }
     return result;
@@ -631,7 +648,7 @@ class SqfliteLocalMediaDb implements LocalMediaDb {
       if (row != null) {
         result.add(row);
       } else {
-        debugPrint('failed to deserialize video playback from row=$rowMap');
+        debugPrint('$runtimeType failed to deserialize video playback from row=$rowMap');
       }
     }
     return result;