+++ test/fake/db.dart
@@ -120,7 +120,7 @@ class FakeAvesDb extends Fake implements LocalMediaDb {
   // dynamic albums
 
   @override
-  Future<void> clearDynamicAlbums() => SynchronousFuture(null);
+  Future<int> clearDynamicAlbums() => SynchronousFuture(0);
 
   @override
   Future<void> addDynamicAlbums(Set<DynamicAlbumRow> rows) => SynchronousFuture(null);